refactor: common 패키지 제거
