+++ tomcat/src/main/java/com/techcourse/handler/StaticResourceRequestHandler.java
@@ -1,8 +1,8 @@
 package com.techcourse.handler;
 
-import com.techcourse.http.common.ContentType;
-import com.techcourse.http.common.HttpCookie;
-import com.techcourse.http.common.HttpVersion;
+import com.techcourse.http.ContentType;
+import com.techcourse.http.HttpCookie;
+import com.techcourse.http.HttpVersion;
 import com.techcourse.http.request.HttpRequest;
 import com.techcourse.http.response.HttpResponse;
 import com.techcourse.http.response.ResponseBody;
