+++ tomcat/src/main/java/com/techcourse/http/HttpMethod.java
@@ -1,4 +1,4 @@
-package com.techcourse.http.common;
+package com.techcourse.http;
 
 import com.techcourse.exception.NotFoundException;
 import java.util.Arrays;
