+++ tomcat/src/main/java/com/techcourse/http/response/HttpStatus.java
@@ -1,4 +1,4 @@
-package com.techcourse.http.common;
+package com.techcourse.http.response;
 
 
 public enum HttpStatus {
