+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -4,9 +4,9 @@ import com.techcourse.exception.NotFoundException;
 import com.techcourse.exception.UncheckedServletException;
 import com.techcourse.handler.RequestHandler;
 import com.techcourse.handler.RequestMapping;
-import com.techcourse.http.common.ContentType;
-import com.techcourse.http.common.HttpCookie;
-import com.techcourse.http.common.HttpVersion;
+import com.techcourse.http.ContentType;
+import com.techcourse.http.HttpCookie;
+import com.techcourse.http.HttpVersion;
 import com.techcourse.http.request.HttpRequest;
 import com.techcourse.http.request.RequestBody;
 import com.techcourse.http.request.RequestHeader;
